Changed option --times-from-file to --lenghts-from-file Changed gene file to alignment file.
git-svn-id: https://svn.vital-it.ch/svn/hp2c/trunk/Codeml_Variants/Fastcodeml@5470 95c5a404-1f88-0410-a6b0-c3f062b6f34a

+++ TreeNode.h
@@ -141,7 +141,7 @@ public:
 	///
 	TreeNode* getChild(unsigned int aIdx) const
 	{
-		if(aIdx >= mChildrenList.size()) return NULL;
+		if(aIdx >= static_cast<unsigned int>(mChildrenList.size())) return NULL;
 		return mChildrenList[aIdx];
 	}
